+++ src/host/alsa/mod.rs
@@ -7,12 +7,9 @@ use ChannelCount;
 use BackendSpecificError;
 use BuildStreamError;
 use DefaultFormatError;
-use Device as DeviceTrait;
 use DeviceNameError;
 use DevicesError;
-use EventLoop as EventLoopTrait;
 use Format;
-use Host as HostTrait;
 use PauseStreamError;
 use PlayStreamError;
 use SampleFormat;
@@ -21,10 +18,10 @@ use SupportedFormatsError;
 use StreamData;
 use StreamDataResult;
 use StreamError;
-use StreamId as StreamIdTrait;
 use SupportedFormat;
 use UnknownTypeInputBuffer;
 use UnknownTypeOutputBuffer;
+use traits::{DeviceTrait, EventLoopTrait, HostTrait, StreamIdTrait};
 
 use std::{cmp, ffi, mem, ptr};
 use std::sync::Mutex;
