+++ src/host/coreaudio/mod.rs
@@ -5,12 +5,9 @@ use ChannelCount;
 use BackendSpecificError;
 use BuildStreamError;
 use DefaultFormatError;
-use Device as DeviceTrait;
 use DeviceNameError;
 use DevicesError;
-use EventLoop as EventLoopTrait;
 use Format;
-use Host as HostTrait;
 use PauseStreamError;
 use PlayStreamError;
 use SupportedFormatsError;
@@ -18,10 +15,10 @@ use SampleFormat;
 use SampleRate;
 use StreamData;
 use StreamDataResult;
-use StreamId as StreamIdTrait;
 use SupportedFormat;
 use UnknownTypeInputBuffer;
 use UnknownTypeOutputBuffer;
+use traits::{DeviceTrait, EventLoopTrait, HostTrait, StreamIdTrait};
 
 use std::ffi::CStr;
 use std::fmt;
