+++ src/host/wasapi/mod.rs
@@ -3,19 +3,16 @@ extern crate winapi;
 use BackendSpecificError;
 use BuildStreamError;
 use DefaultFormatError;
-use Device as DeviceTrait;
 use DeviceNameError;
 use DevicesError;
-use EventLoop as EventLoopTrait;
 use Format;
-use Host as HostTrait;
 use PlayStreamError;
 use PauseStreamError;
 use StreamDataResult;
-use StreamId as StreamIdTrait;
 use SupportedFormatsError;
 use self::winapi::um::winnt::HRESULT;
 use std::io::Error as IoError;
+use traits::{DeviceTrait, EventLoopTrait, HostTrait, StreamIdTrait};
 pub use self::device::{Device, Devices, SupportedInputFormats, SupportedOutputFormats, default_input_device, default_output_device};
 pub use self::stream::{EventLoop, StreamId};
